requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
jinja2==3.1.4
python-dotenv==1.0.1
yfinance==0.2.40
feedparser==6.0.11
nltk==3.9.1
google-generativeai==0.8.3
aiofiles==24.1.0
